// generators/src/types.ts
export type KarabinerRules = {
    description?: string;
    manipulators?: Manipulator[];
};

export type Manipulator = {
    description?: string;
    type: "basic";
    from: From;
    to?: To[];
    to_after_key_up?: To[];
    to_if_alone?: To[];
    parameters?: Parameters;
    conditions?: Conditions[];
    to_delayed_action?: {
        to_if_invoked: [
            { set_variable: { name: string; value: number } },
            { key_code: KeyCode },
        ];
    };
};

export type Parameters = Partial<{
    "basic.simultaneous_threshold_milliseconds": number;
    "basic.to_delayed_action_delay_milliseconds": number;
}>;

type Conditions =
    | FrontMostApplicationCondition
    | DeviceCondition
    | KeybaordTypeCondition
    | InputSourceCondition
    | VaribaleCondition
    | EventChangedCondition;

type FrontMostApplicationCondition = {
    type: "frontmost_application_if" | "frontmost_application_unless";
    bundle_identifiers?: string[];
    file_paths?: string[];
    description?: string;
};

type DeviceCondition = {
    type:
        | "device_if"
        | "device_unless"
        | "device_exists_if"
        | "device_exists_unless";
    identifiers: Identifiers;
    description?: string;
};

type Identifiers = {
    vendor_id?: number;
    product_id?: number;
    location_id?: number;
    is_keyboard?: boolean;
    is_pointing_device?: boolean;
    is_touch_bar?: boolean;
    is_built_in_keyboard?: boolean;
};

type KeybaordTypeCondition = {
    type: "keyboard_type_if" | "keyboard_type_unless";
    keyboard_types: string[];
    description?: string;
};

type InputSourceCondition = {
    type: "input_source_if" | "input_source_unless";
    input_sources: InputSource[];
    description?: string;
};

type InputSource = {
    language?: string;
    input_source_id?: string;
    input_mode_id?: string;
};

type VaribaleCondition = {
    type: "variable_if" | "variable_unless";
    name: string | number | boolean;
    value: string | number;
    description?: string;
};

type EventChangedCondition = {
    type: "event_changed_if" | "event_changed_unless";
    value: boolean;
    description?: string;
};

export type SimultaneousFrom = {
    key_code: KeyCode;
};

export type SimultaneousOptions = {
    key_down_order?: "insensitive" | "strict" | "strict_inverse";
    detect_key_down_uninterruptedly?: boolean;
};

export type From = {
    key_code?: KeyCode;
    simultaneous?: SimultaneousFrom[];
    simultaneous_options?: SimultaneousOptions;
    modifiers?: Modifiers;
};

export type Modifiers = {
    optional?: string[];
    mandatory?: string[];
};

export type To = {
    key_code?: KeyCode;
    modifiers?: KeyCode[];
    shell_command?: string;
    set_variable?: {
        name: string;
        value: boolean | number | string;
    };
    mouse_key?: MouseKey;
    pointing_button?: string;
    /**
     * Power Management plugin
     * @example: sleep system
     * @see: {@link https://karabiner-elements.pqrs.org/docs/json/complex-modifications-manipulator-definition/to/software_function/iokit_power_management_sleep_system/}
     */
    software_function?: SoftwareFunction;
};

export interface MouseKey {
    y?: number;
    x?: number;
    speed_multiplier?: number;
    vertical_wheel?: number;
    horizontal_wheel?: number;
}

export interface SoftwareFunction {
    iokit_power_management_sleep_system?: {};
}
// ...
export type RectangleActions =
    | "left-half"
    | "right-half"
    | "maximize"
    | "maximize-height"
    | "previous-display"
    | "next-display"
    | "larger"
    | "smaller"
    | "bottom-half"
    | "top-half"
    | "center"
    | "bottom-left"
    | "bottom-right"
    | "top-left"
    | "top-right"
    | "restore"
    | "first-third"
    | "first-two-thirds"
    | "center-third"
    | "last-two-thirds"
    | "last-third"
    | "move-left"
    | "move-right"
    | "move-up"
    | "move-down"
    | "almost-maximize"
    | "fill-left"
    | "fill-right"
    | "center-half"
    | "first-fourth"
    | "second-fourth"
    | "third-fourth"
    | "last-fourth"
    | "top-left-sixth"
    | "top-center-sixth"
    | "top-right-sixth"
    | "bottom-left-sixth"
    | "bottom-center-sixth"
    | "bottom-right-sixth"
    | "first-sixth"
    | "last-sixth"
    | "fullscreen"
    | "close"
    | "minimize"
    | "quit-app"
    | "hide-app"
    | "cascade-all"
    | "cascade-app"
    | "tile-2x2"
    | "tile-2x3"
    | "reveal-desktop-edge"
    | "app-next-display"
    | "app-prev-display"
    | "app-left-half"
    | "app-right-half"
    | "first-three-fourths"
    | "last-three-fourths"
    | "top-left-ninth"
    | "top-center-ninth"
    | "top-right-ninth"
    | "middle-left-ninth"
    | "middle-center-ninth"
    | "middle-right-ninth"
    | "bottom-left-ninth"
    | "bottom-center-ninth"
    | "bottom-right-ninth"
    | "top-left-third"
    | "top-right-third"
    | "bottom-left-third"
    | "bottom-right-third"
    | "top-left-eighth"
    | "top-center-left-eighth"
    | "top-center-right-eighth"
    | "top-right-eighth"
    | "bottom-left-eighth"
    | "bottom-center-left-eighth"
    | "bottom-center-right-eighth"
    | "bottom-right-eighth"
    | "center-two-thirds"
    | "fill-bottom-left"
    | "fill-bottom-right"
    | "fill-top-left"
    | "fill-top-right"
    | "last"
    | "next-space"
    | "nudge-left"
    | "nudge-right"
    | "nudge-up"
    | "nudge-down"
    | "prev-space"
    | "snap-bottom-left"
    | "snap-bottom-right"
    | "snap-top-left"
    | "snap-top-right"
    | "upper-center"
    | "next-display-ratio"
    | "prev-display-ratio"
    | "stash-left"
    | "stash-right"
    | "stash-up"
    | "stash-down"
    | "unstash"
    | "cycle-stashed"
    | "toggle-stashed"
    | "unstash-all"
    | "stash-all"
    | "stash-all-but-front"
    | "reflow-pin"
    | "next-space"
    | "prev-space";